Update test cases to do db-updates/inserts and more validations (also removed dupplicated tests)
